+++ view/src/commandHandlers/gHandlers/g03Handler.java
@@ -14,7 +14,7 @@ public class g03Handler {
 		return false;
 	}
 	
-	public static void execute(int x, int y, int i, int j) {
+	public static void execute(double x, double y, double i, double j) {
 		if (test() == false) {
 			return;
 		}
